requirements.txt
gunicorn
flask
flask-sqlalchemy
flask-migrate
wtforms
flask-bcrypt
flask-admin
python-dotenv
pytest-flask
flask_cors
pandas
numpy
matplotlib
seaborn
dataframe_image
sklearn